Use internal ELB for Elasticsearch cluster
Internal ELB for Elasticsearch allows using SourceSecurityGroup to open access
to Elasticsearch ELB from search-api instances. (Using SourceSecurityGroup
allows access to private EC2 IPs only, and so doesn't work with internet-facing
ELB).

Internal ELB resolves to internal instance IPs and is unavailable outside
the VPC, so for dev access we open ports on Elasticsearch instances themselves.

Internal ELB has to be configured with subnets instead of AZs, and since we're
not managing subnets with CloudFormation right now and there doesn't seem to be
a way to retrieve the list of default subnets automatically, we're adding a
list of default subnets to environment vars file for now.