Do not replace existing EB application versions
Trying to create a version with an existing verison name
does not replace the existing one.

The source package uploaded to S3 is replaced, but the currently deployed
version does not use the package from the S3 bucket, so the change
does not apply unless the deployed version is changed beforehand.

Disabling S3 uploads if the S3 key exists to avoid this.