Add promote command
If called with a stage of 'staging' it will attempt to load the release
from 'preview'.
If called from 'production' it will ignore the 'release_name' and release
whatever is currently in 'staging'.

Creating an instance of Deploy was moved into StackPlan as when needed
it's easier to create from there.