Do not require user.yml to be present
It is not needed for releases from the jenkins machine.
This change prevents the user.yml being added to the vars files if it
does not exist.
It also changes the built stack parameters to be lazy so that some can
be evaluated without all being evaluated. This is needed for the release
command so that parameters not required for release do not have to be
evaluated.