Do not write version label if loading from file
When opened in append mode writing to the zip file adds more records to
it even if a record of the same name exists. So, if we're already
dealing with an archive then if we set the version label again the zip
cannot be unzipped without human intervention, causing the deploy to
fail.