Add loaddata command
Downloads a database dump from S3 and runs it on the environment
RDS instance.

Runs Elasticsearch import script from the digitalmarketplace-api
repository with environment API and Elasticsearch as arguments.

A lot of this will change once we have write endpoints on api and
search-api that would allow loading data throught them instead of
using the databases directly.