Add dmaws.build tests and related fixtures
Adds an optional fixture for run_cmd that allows setting the output
and checking the calls easier than using subprocess.

Since most of the build functions have very little logic some tests
just verify that relevant mocks are in place and the functions aren't
trying to access the file system/subprocesses.

For more complicated functions we test that the git commands performed
are what we expect.

Testing with a real git repo would actually test that the git commands
are correct, but this would make the tests much slower and maintaining
a repo that would allow tests for different conditions might be
problematic. Mocks give us reasonable confidence to change code in the
future, as long as we know what the git commands involved actually do.