Add Beanstalk configoptions for rolling deploy
This ensures that deploys only go out to one instance at a time.
See:
http://docs.aws.amazon.com/elasticbeanstalk/latest/dg/using-features.rolling-version-deploy.html

I have tested this by increasing the number of instances and running a
deploy while providing constant load with vegeta. When the app is
experiencing no load the deploys appear to overlap. However, when there
is load and these configuration options are set there are no 504s. When
these options are not set there are a small number of 504s (3%).