Add nginx ansible local provision with ASG user-data script
CloudFormation sets user data on the ASG launch configuration using
the template parameter values.

The user data is executed as a script by cloud-init when ASG launches
an instance. The script will then run the relevant tasks to set up
the instance config files.

Using user-data allows us to keep per-environment and per-stage
variables out of the AMI and therefore use the same AMI for every
environment, which means that we're running the same software, only
the configuration files are different.

The user data script relies on the ansible playbook and roles
being available on the instance, so they should be copied during
image creation.