Add packer template for nginx AMI
Creates an EBS-backed image and installs nginx using image-setup
tasks in ansible nginx role.

ansible-local provisioner copies the ansible playbook and role files
and setting the explicit staging_directory makes sure the files are
stored inside the image. That way, they are available for the user
data scripts to run the instance-config tasks.