Handle RateLimit exceptions in Cloudformation.describe_stack
When running a delete/info command it's possible to make a large
number of requests with small delays, which leads to 400 rate limit
responses from the CF API. BotoServerError try block was catching
these exceptions and handling them as the "stack doesn't exist",
which is wrong.

Instead, we look for the explicit "doesn't exist" message to recognize
missing stacks, retry the rate limited requests after a delay, and
log and reraise everything else.