Set DNS resolver and variables for proxy_pass
When proxy_pass is used with a predefined URL nginx will resolve
the host during startup and will cache the IP address indefinitely.

This is a problem when proxying requests to AWS ELBs, since ELB IPs
change without notice.

To force nginx to re-resolve DNS proxy_pass destination must be set
to a variable and resolver must be defined in server configuration.

Nginx can't use the nameserver IP from system resolv.conf, so we're
passing it in as another variable.

Since we set resolver in nginx.conf and the variable is only set
in cloud-init stage nginx.conf is now run during instance-config.