Increase the number of app instances on production
See: https://www.pivotaltracker.com/story/show/92327910

Increase the number of EC2 instances per app in production to 3. This
matches the number of availability zones for the EU west region. The
assumption is that elastic beanstalk will automatically create the
instances across the availability zones. This will be tested when this
goes out.