Add robots.txt to block /admin and /suppliers indexing
[#99500206](https://www.pivotaltracker.com/story/show/99500206)

Adds robots.txt file served from the root of www. Disallows any
URLs starting with `/admin` or `/suppliers`, so that both root
pages and all other pages of admin and supplier apps are not
indexed.

Since suppliers-guide was moved to `/g-cloud/suppliers-guide`,
there's no need to add an allow rule for it.