Enable connection draining for nginx ELB
Connection draining will keep instance InService for the given
time to allow existing requests to finish. This way nginx AMI
updates could be applied without interrupting or failing any
requests.

The timeout is set to 300s which should be the same as the default
value. This is the limit that ELB will wait for requests to finish
before closing any remaining ones and removing the instance.

This is the same setting we've enabled to fix failing requests on
application deploys.