Set nginx ASG update pause
ELB-based health checks for ASG instances are not used for adding
the instance to the pool, only for removing the current instances.
Instances are added to the pool as soon as the EC2 check finishes,
and health check grace period only affects the ELB health checks
that would remove the InService instance.

So the ELB health check doesn't add the necessary delay required
for nginx instances to finish setup when replacing instances during
the launch configuration update.

Adding pause time forces Cloudformation to wait between relaunching
instances. This would work better if we had `"WaitOnResourceSignals"`
set up, which would continue the update once the instance reported
as ready, but for now the 10 minute delay should be enough for nginx
to come up without increasing the update time too much.