Add a CloudWatch metric for Nginx requestTime
This will create a CloudWatch metric filter that takes the requestTime
from Nginx JSON logs and puts it into a CloudWatch metric. In preview
the metric will have a namespace preview-preview/Nginx and a name of
requestTime. This doesn't match metric naming convention we've used to
date because we don't have access to the ElasticBeanstalk environment
name. This may actually turn out to be a more flexible naming
convention.