Unset DM-Request-ID on initial requests
The way that we pass the DM-Request-ID around services is through the
HTTP header. As a result, a client could set the DM-Request-ID
themselves from outside. This strips that header on the way in so that
one is generated internally.

There is no unset header directive, instead an empty value prevents the
header being passed on to the proxied server.

http://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_set_header