Move 5xx metric filters and alarms to dmaws
Move the 5xx metric filters and alarms into dmaws from the applications
.ebextensions.

A short version of the application name has been added to the
`DM_METRICS_NAMESPACE` to match the namespace used for Nginx logs and be
easier to search for than making the metric names themselves really
long. This is the AWS recommended way of organising metrics. For
example, the metrics namespace for the supplier app in preview would be
`preview-preview/supplier-frontend`.

Metric filters and alarms have been pulled out into macros to avoid
mistakes in boilerplate such as log group name, alarm actions and
namespace.

The metric filter macro accepts a sub-part of the patter which it wraps
and adds the application filter to automatically. For example, giving
the metric filter `$.foo = bar` to the macro will result in a metric
filter of `{ $.application = api && ($.foo = bar) }` on the api app.

Alarm descriptions are implemented in separate templates so that they
can be easily managed and more easily made uniform (eg. extending from a
base template).

After discussion we have decided not to port over the 4xx and related
alarms unless they can be shown to be useful. Instead favouring more
specific alarms where needed.