Write a KEY=VALUE property file using release command output
Uses ctx.out to write KEY=VALUE pairs to stdout, which is redirected
by the jenkins script to a property file. Stdout is now reserved for
this output format, all other unstructured command output should be
written to stderr by `ctx.log`.

This allows us to communicate release information back to the caller
environment (e.g. Jenkins) to use in the following build steps and
notifications.