Allow promoting preview to staging by setting empty release-name
If release-name is not set when calling release to staging current
preview release version will be used. This should act similar to the
current production releases.

"release-name" is kept as a parameter and will be used when set. Staging
is the only stage that allows an explicit release name right now, so
this is required for rollbacks (rollbacks should start with staging and
can be promoted to production, preview can not be rolled back right
now).

"release-name" check is changed to `not release_name` to avoid adding
logic to the jenkins script, which will always pass `--release-name=`
for staging, but now an empty value is allowed.