Increase instance size and max limit for API and supplier app
Supplier app reached the autoscaling limit of 5 instances.
Since we expect an increase in requests for the last days of the
G-Cloud 7 service submission, we raise the max limit and size of
EC2 instances for supplier frontend and Data API apps.

We can reduce these limits to the previous value  after the service
submission is closed.