Add method to prune old Beanstalk app versions
There is a limit to the number of Beanstalk application versions that
you can have, this limit is 500 per account and AWS refuses to increase
it stating that we should be deleting old versions.

This adds a method to prune old Beanstalk application versions by
listing them and deleting any above the number to keep.