Add method to restore an instance from snapshot
This is very geared towards our needs for syncing data. The snapshot is
created with a hard coded micro instance size and single AZ. Once
created the instance is modified to update the same VPC security groups.
This has to be done with a modify because the option is not available
when calling restore.

This change also extracts waiting for an RDS entity to be available as
it's used a few times now.