Move variable file loading into Context
This change moves the code that builds a list of variables files to load
from the cli module into the Context object so that Context objects for
other environments can more easily be created within a command.

This is needed for migrating data because we need to load variables
associated with the target environment.