Add methods to allow/revoke access to RDS instance
This adds a number of methods to the RDS class to provide access
control to an RDS instance.

allow_access_to_instance first revokes access to the instance to avoid
having to inspect the security group and figure out if the access we
require is provided by it. It's simpler just to remove and recreate it.

When modifying an RDS instance the instance does not immediately go into
the 'modifying' state, but also, they sometimes change so quickly that
the returned state is already back to 'available'. Therefore when we
modify the instance we first wait for the instance to stop having the
'available' state, but only for a limited number of iterations, then we
wait for it to have the 'available' state.

When revoking access if the required security group is not on the
instance then it's a noop. If the required security group is the only
security group on the instance then it is replaced with the default
security group as RDS ignore the parameter if it's empty.