Provide SHARED_EMAIL_KEY evironment variable
DM_SHARED_EMAIL_KEY is set to SHARED_EMAIL_KEY in the base Config class
which has the same problems as described in the previous commit.
Ordinarily I would suggest changing the apps to use DM_SHARED_EMAIL_KEY,
however none of the other email related environment variables have the
DM_ prefix and it is used directly from the app config in dmutils.