Update RDS database to PostgreSQL 9.4
Changes the database engine version to 9.4. Major upgrades can only
be done one at a time, so there's no way to move from 9.3 to 9.5 in
one step. "allowMajorUpgrades" has to be set to true to apply the
change.

Despite the description in the [AWS docs][1], there doesn't seem to
be a way to change the DBParameterGroup Family parameter. AWS sets
the resource status to 'UPDATE_COMPLETE', but the parameter group
still has the old database version. And there doesn't seem to be
a way to update the parameter group family using AWS CLI, so it's
likely an error in the CloudFormation docs.

The updated database version however requires a matching parameter
group family, so we rename the parameter group resource to force
CloudFormation to replace the old group with a new one. This runs
the expected update, removing the old group once the DB instance
has finished updating, but requires a manual change to the template
each time.

[1]: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-rds-dbparametergroup.html#cfn-rds-dbparametergroup-family