Remove dependency on monitoring from database stack
Database stack doesn't reference the monitoring output, so there's
no reason to have an explicit dependency. The only connection between
RDS and monitoring is the notification when a DB instance is deleted,
but there's little value in them for standalone databases.

Monitoring is still required for apps, so there's no risk of
accidentally skipping the stack update.