Add AWS Lambda CloudFormation template and stacks
There are two ways to create a Lambda function using CloudFormation:
* by inlining the code into the template directly, which is limited
  to 4096 bytes
* from an S3 object containing a ZIP archive with the function code

Since the inline size is quite restrictive we're going with the S3
route. The stacks will create an S3 bucket and then attempt to
create a Lambda function using the `S3Key` object that should be
stored in that bucket. Since there's no way to upload a file to S3
using CloudFormation that file has to be uploaded manually (after
the bucket stack was created, but before the Lambda one) when the
stack is run for the first time.

Rerunning the stack if the S3 file was changed won't update the
Lambda code, since CloudFormation considers the template to be
unchanged. There's a way to add S3 object version to the resource
(or use manual versioning by changing the file name for each version)
forcing CloudFormation to apply the update, but that requires changing
the variable for each new function version (similar to how packer AMI
flow works right now). Instead, since code archive creation will require
additional steps anyway, we're going to add a command to release and
apply a new lambda version.