Add lambda-release dmaws command to push lambda code updates
Since there's no convenient way to upload Lambda function code with
CloudFormation we need a separate command to create the code archive,
upload it to S3 and update the Lambda.

`lambda-release` command takes the name of the lambda function stack
and:
1. Looks for the function code in `lambdas/<lambda_name>`.
   `lambda_name` is the name of the stack with `"_lambda"` part removed.
   The folder should contain a main.js file (since this is what our
   runtime and handler parameters in aws_lambda template are set to).
2. Create the environment file with all `EnvVar` parameters stored
   in JSON.
3. Archive the code and upload it to `S3Bucket/S3Key`
4. Update the Lambda function code with the new file from S3

Environment variables are taken from the stack parameters matching the
`EnvVar` naming convention. Parameter names are rewritten using `param_to_env`,
similar to how we do it with ElasticBeanstalk app environment variables. The
file is stored as `'env'` and added to the code package, so that it can be
loaded by the function code.

Since all CloudFormation parameters need to be declared by the template,
even though they're not used by any of the resources, we have to generate
the EnvVar declarations using Jinja, same as in the app_base template.

We could update function code directly instead of uploading to S3 first,
but this approach makes it possible to recreate CloudFormation stack
without losing the function code. Since CloudFormation references the
same S3 object it will recreate the stack with the latest function code.
The downside of this approach is that we need to upload a file to S3
before the first lambda stack run. The alternative would be to use an
empty function code block in the CloudFormation template and then upload
the actual code using the new command.