Add log streaming from CloudWatch to Lambda function
Creates a log subscription for json log group. There can be only
one subscription per log group, which we set up with an empty
filter, so that all log streams are sent to the lambda function,
which will then store them in Elasticsearch. This is not ideal,
since our log group contains both apache and application logs
and with one subscription the only way to split them is inside
the lambda function itself. And there's no way to set up separate
functions for more granular filters (eg notifying on exceptions).