Make Kibana Elasticsearch cluster only accessible from Jenkins
Since there's no way to set up user authentication on AWS ES service
we make the Elasticsearch server accessible only from the Jenkins IP
and set up a proxy with authorization on the Jenkins instance. This
way ES is protected both by an IP restriction applied to the Jenkins
nginx port and the authorization set up in nginx.