Add a stack to open Jenkins port for letsencrypt checks
When verifying the domain ownership letsencrypt CA needs to be able
to make a request to the domain server. Since the source IP address
can change often we open the HTTP port for all IPs and change nginx
config on Jenkins instance to only listen on 443, which means our
proxy is still restricted to our developer IPs and port 80 is free
to be used by letsencrypt standalone server when required.

The stack can be safely removed when unsused since it's only needed
for the verification when requesting the certificate.