Restore staging database from a snapshot
Unintentionally reset the staging database over the weekend, so
restoring from a snapshot taken this weekend to see if we get our
data back.
Our hypothesis is that we won't, but
- at least we'll know for sure
- it's an opportunity for me to try out our backup process, if
you want to call it a process

Also toyed with the idea of increasing the number of days we keep
around our backups, but that doesn't really seem like the problem
that we have.