Reduce lambda function timeout to 5s
When Kibana server is under heavy load it stops responding to lambda
requests, which makes the function wait until it's interrupted by
a timeout. Higher timeout value means the function will wait longer,
which increases the AWS bill.

Most successfull requests to Elasticsearch should complete within 5s,
so setting that as the timeout value. It could be adjusted if the
lambda monitoring shows a lot of functions being interrupted.