Move stage argument after the command name
There doesn't seem to be a way to organize commands with and without
arguments before the command name in a single CLI, so in order to
support the helper commands I'm moving the STAGE argument after the
command name.

This also removes ENVIRONMENT argument. We haven't used environment
names that don't match the stage name until now, and we don't have
any plans to at the moment, so I'm hiding the separate argument and
making sure that environment is always the same as stage.