Add S3 logs bucket and enable S3 access logging
Enables access logging for all S3 buckets. Logs are stored in a
separate logs bucket for each environment and logs from buckets
are stored in separate folders. Log bucket access logs are stored
in the same bucket (which thankfully doesn't create an infinite
feedback loop).

Enabling logging was recommended by ITHC report. I'm not sure what
the cost implications are since it seems to create a lot of small
files, and S3 could get quite expensive for LIST/PUT operations,
so we'll have to keep an eye on it.