Our minimal viable declaration has increased
During the signing flow, we expect two more fields to be in the
declaration:

- the organistaion name
- the email address of the primary contact

A real email will be sent to this email address on staging, so
let's not use a real address.

Amazingly, at least one of our supplier names contain \"double
quotes\", so I've needed to put in a replacement function that
strips them out.