Delete database instance before creating new snapshot
The old code would presume no database instance existed and
just start trying to create a new one with a hardcoded id.
However, if an old instance was lying around (for example, if a
previous migration had failed somewhere), no further migrations
could be performed until the old instance was successfully deleted.

Now, we check first for an existing database and remove one if we
find it.

Note that there are a few instances of the "delete if found"
behaviour in this script, but I couldn't think of a good way to
make the behaviour generic.  If this is a priority, I can make
it happen, but I assumed it wasn't.