Upgrade bcrypt
Running the migration was causing some warning in the output:
something about not explicitly casting the right thing for bcrypt.
(seen here: https://github.com/pyca/bcrypt/issues/82#issuecomment-247816431)

The solution seems to be to upgrade bcrypt, which _looks_ scary
because it goes up a major version but actually the changelog
doesn't seem too serious.

[Changelog](https://github.com/pyca/bcrypt#changelog)