Open dev access to target stage
The way I understand it, we've been creating a security group
for our new `exportdata` instance (which gives Jenkins
permission to connect to the database) and then attaching it
to the target stage as well so that Jenkins can access the
target database (ie, to move data into it).

Reusing our new security group this way has occasionally meant
that -- if the script fails to clean up everything -- there is
open access to our target database which isn't cleaned up
automatically.

Opening `database_dev_access` on the target stage gives jenkins
access and it also gets closed nightly, so there's less potential
for security holes.