Switch from security group preview/staging IP restriction to nginx
We've added elasticsearch proxy config for PaaS to our existing
nginx cluster. Which means nginx is now responsible for several
things that have different access restrictions:

* preview and staging www. environments should be locked to the
  AH IP range
* elasticsearch should be available from the PaaS search-api
  instances

But since we're using AWS security group rules to restrict access
there's no way for us to define different rules based on the
request Host header.

So instead we move the IP check from the security group to the
nginx config.

We can revert this back to the security group which provides a
clearer restriction and is easier to update once we no longer
need the elasticsearch proxy.