Force push release tags in EB deployment jobs
We're running PaaS deployments in parallel with existing EB ones at the
moment, and they're creating the same release-* tags. Which means we can
no longer depend on the EB application version existing when the tag
exists. Since we don't have a way to check that EB version exists,
we ignore any errors instead and force the tag push.