Add Make rule to populate PaaS db from dump
We need to set up an ssh tunnel via the api app to allow importing of
data into the PaaS db's from a local dump.

The `PAAS_SPACE` needs to be provided as a parameter which is checked
against the currently targeted cf stage to prevent accidental writing to
the wrong db.