Enable sticky sessions for nginx elb
We've enabled sticky sessions for the nginx ELB on the PaaS environment
to make sure that users get a consistent experience when we move to
maintenance mode.

If, for example, we're moving form live to maintenance mode, during the
transition the new maintenance nginx instances will be added to the
autoscaling group with the old ones. This means for a short period of
time a user could initially hit a live instance and get the html for
the page, but then when the browser makes further requests for css/js
it could get a 503.

Weirdness.

Enabling sticky sessions means that if they hit a live nginx instance,
further requests for all css/js will also hit that instance. Similarly,
as soon as they hit a maintenance instance they will only see the
maintenance page, even on a refresh of the page.