Change paas-deploy process to not require manual rollback
Previous release process was renaming the app to `{app}-rollback`. If
the cf push failed to bring up the new application version then the
release failed, leaving the rollback application, which means that
on the next release attempt the renaming fails since `{app}-rollback`
is already taken.

The new process instead pushes the new app as `{app}-release`. If
the release succeeds then the old application is unmapped from the
route and deleted and `{app}-release` is renamed to `{app}`.

This change also means that if an app is accidentally pushed using
`cf push` and the generated manifest it won't bring down the production
application during release, but will create a separate `{app}-release`
application instead.

If the release fails then both `{app}` and `{app}-release` will exist,
and the next release will push to `{app}-release` as usual.

This also fixes the need to manually create an app before attempting
the release for the first time.

We're removing the scaling step from the release process since
autoscaling is not set up on PaaS yet so the manifests should serve
as the main source of truth for the desired instance counts.

This also introduces a delay between removing the traffic from old
application version and shutting down to allow the instances to push
any outstanding logs to CloudWatch.