Update paas-manifest command to load per-app manifests
Removes template file option in favour of the app-specific manifests.

Changes the way variables are set for the manifest template. Once
the relevant variable files are loaded and merged the command will
update the top-level namespace with the values from the application.

This allows us to use the same base manifest template referencing
generic top-level variables (eg `subdomain`, `instances`, `path`)
that are overridden by application-specific values. Previously
this was accomplished by using `stacks.yml` as the middle layer.

The template variable change means that we can run into issues if
we use clashing variable names accidentally, but at the same time
it allows us to set common values for all applications. Eg:

```
instances: 3

api:
  instances: 5
```

sets instance counts to 3 for all applications, but since the context
will be updated with the `api` values the api manifest will only see
`instances: 5` value.