Change production domain to service.gov.uk
The production state file has been updated so that the nginx aws route53
zone is now the original digitalmarketplace.service.gov.uk zone. This is
so that when we make the switch from production.marketplace.team, we
don't need to recreate the zone file and lose the name server records.

The domain is updated here to match the new zone file. The original
subdomain records for api, assets etc will need to be deleted, so that
Terraform can recreate them.