Stop db migration app before running the task
With a finite sleep command db migration app keeps restarting every
N hours.

Since the CF tasks spin up a new container for the task we don't
need any application instances to keep running once the code is
pushed, we're only using the db migration app as the source of
configuration and container data.

So we should be able to stop it before running the task.