Replace dmaws commands with standalone scripts
Most of dmaws modules are no longer used since we've switched to
using Terraform and PaaS. There's only one `dmaws` command that's
currently being used - generate PaaS manifest file.

Since a lot of dmaws functionality was tied to managing stack files
the easiest way to removed unused code and keep the PaaS manifest
generation is to pull it into a separate script that uses some of
the dmaws helper functions.

This allows us to remove all AWS-related modules and tests, but keep
the variable loading logic.

I'm also keeping dmaws/rds.py for now. While it's no longer being
used it contains the SQL commands we use to clean up the database
between production and staging and we don't know yet how different
the migration process will be on PaaS.