Fix log streaming errors caused by nested objects in log fields
Nested keys in log message fields are causing issues with Kibana,
especially when the field type changes between messages. Eg api_error
can be both a string message or a dictionary, which isn't valid in
Elasticsearch index mapping.

Since we don't use the nested keys for searching and filtering and
and having them mapped creates issues with log streaming we can
serialize them to JSON strings in the streaming Lambda function.

All nested objects and arrays are serialized.