Split list of log groups and create subscriptions separately
Previously we were creating log group/lambda subscriptions by giving the
log-streaming module a list of all log group names. This list was made
of the main nginx json log group, app nginx logs and app application
logs, all concatenated together.

This is fine until a new app (briefs frontend) was added. The
subscriptions were created by iterating over the full list of names;
adding a new app in the middle shifted the list so, Terraform wanted to
recreate about half of the subscriptions. This would result in some
period of time when logs would not be streamed to Kibana.

By splitting the log groups names into two separate groups
(`application_log_groups` and `nginx_log_groups`) we can now append a
new app to the end of the list and not disturb the order of the existing
apps, and so not forcing recreation.

The nginx json logs and the app json logs have been left together as
they can both use the same filter, and there's not much chance of adding
any more groups for the main nginx.

The filter has been split between the two new subscriptions types.
Previously it could be applied to nginx logs and application logs. With
the separate log groups it can be much more streamlined.