Set briefs subdomain to `dm`
The briefs subdomain was set to `dm-briefs` so the new app could be
released without receiving traffic straight away so we could ensure it
worked properly without breaking the buyer frontend.

Now that we've tested it, we can change the subdomain here so subsequent
releases are assigned the correct subdomain.