Add Makefile rules to create and check a snapshot service
We can make a postgres service using the data in the latest db snapshot
that PaaS take for us. From this we can take a dump which we can store
offsite for security.

It takes a while to spin up, so it's useful to have a way to check the
status of the new service, so that scripts can wait for the service to
be ready before continuing.