Create app for creating and uploading dump
To access the snapshot service we created, we need an app to connect to
it. This sets up that app with a connection and then makes it sleep.

We will use a cloud foundry task to run a script on it which will do the
work of creating and uploading the dump.

There is an issue with cloudfoundry and the snapshot db services. For
now to test, I'm connecting to the live db. When the issue is resolved
we'll revert to a snapshot db to avoid extra load on the live db.