Add make rules for deploying, checking and cleaning the snapshot app
To allow jenkins to automate the backup process, we have make rules to
deploy, check and destroy the new snapshot app.

The snapshot app is injected with an extra env variable, which contains
the data it needs to upload the dump to s3 via a signed URL. This data
will change each time the backup process is run, and will be generated
by a script run on Jenkins.

The deployed app then runs a script via a task to actually do the work.

The rule to check the snaapshot app just returns it status, it's used
for waiting for the app to be ready before a script can continue.

The cleanup rule destroys the extra service as well as the snapshot app.