Add script to generate a signed url for upload
You can create signed url's for allowing users to upload things to your
s3 buckets without having their own credentials. Boto can manage this
for you.

This script will be run on the Jenkins box, which is in the same account
as the bucket we're going to use to store the dumps. This makes
permissions easy.

This output of this will then be injected into the snapshot-app so it
can be used to upload the dump to s3.