Add script to upload dump to s3
This is a simple python script which uses the signed url previously
generated to upload the encrypted dump to s3 using the requets library.

If the response from s3 is not successful, an error is raised and we
print the error message. Also sets the exit code to be not zero.