Add expiration and filename to s3 post url data
Adding an expiration time in seconds makes the URL more secure. Here
it's set to 10 minutes.

The key is what the file will actually be called in s3. This sets it to
a timestamp of year, month, day, hour, minute.

I've also pulled out the bucket name as a variable so this script could
be used for other buckets.

Added docopt and a docstring to explain how to use.